Remove checks for predetermined archs
Now any arch is considered valid.
When checking for arch dependant variables such as `depends_x86_64`
we check against the archs that the specific SRCINFO supports,
instead of a predetermined list. This means a PKGBUILD that
only supports x86_64 should error if it contains `depends_armv7h`.

Note this will cause a failiue if `depends_x86_64` comes before `arch` in
a SRCINFO but makepkg will never generate it this way so I'm considering
that case invalid.