Add Satisfies() and Restrict() for dep checking
The Dependency type had all the info for dep checking but never had any
way to do so. With slt and sgl being private it made it impossible to do
the checking from the outside.

Satisfies() checks whether a version string falls inside the range
specified by a Dependency.

Restrict() merges two Dependencies together into a new Dependency in
which both of their restrictions are met. If both restrictions can not
be met MaxVer and MinVer will end up overlapping so that using that
Dependency in Satisfies() will end up returining false against every
version.