Add Dependency.String()
Real Dependency strings can only have 1 restriction at a time so
"foo>1<2" would be invalid. For this reason the string function returns
in the format "foo>1 foo<2", it could instead return (string, string)
rather than one string thats space seperated but for the sake of
complying with go's .Sring() conventions it is done this way.