Allow lowercase letters in pkgname
Although Arch has its own packaging guidelines such as package names
being lower case only. This is not a rule set by makepkg itself which is
what I believe we should be following.

Having too restrictive checks have caused problems again and again and
each restriction is slowly being lifted. I mentioned this in #11.

The point of this patch though is not just because we should allow it.
But because it causes real problems with packages such as xorg-server,
which has uppercase provides.