Add test case for PR #25
Add PKGBUILD and SRCINFO from AUR package 'bash-snippets' (SRCINFO seems
to have broken indentation). It is not a Core package however, so maybe
this test case should be moved somewhere away from the script testing
Core and SRCINFOs, or the the test script should be modified not to say
it's testing Core packages only.